+++ packages/components/input/__tests__/input.test.tsx
@@ -37,6 +37,22 @@ describe("Input", () => {
     expect(container.querySelector("input")).toHaveAttribute("disabled");
   });
 
+  it("should disable the clear button when isDisabled", () => {
+    const {getByRole} = render(<Input isClearable isDisabled label="test input" />);
+
+    const clearButton = getByRole("button");
+
+    expect(clearButton).toBeDisabled();
+  });
+
+  it("should not allow clear button to be focusable", () => {
+    const {getByRole} = render(<Input isClearable label="test input" />);
+
+    const clearButton = getByRole("button");
+
+    expect(clearButton).toHaveAttribute("tabIndex", "-1");
+  });
+
   it("should have required attribute when isRequired with native validationBehavior", () => {
     const {container} = render(<Input isRequired label="test input" validationBehavior="native" />);
 
@@ -126,26 +142,6 @@ describe("Input", () => {
     expect(ref.current?.value)?.toBe(value);
   });
 
-  it("setting ref should sync the internal value", () => {
-    const ref = React.createRef<HTMLInputElement>();
-
-    const {container} = render(<Input ref={ref} type="text" />);
-
-    if (!ref.current) {
-      throw new Error("ref is null");
-    }
-
-    ref.current!.value = "value";
-
-    const input = container.querySelector("input")!;
-
-    input.focus();
-
-    const internalValue = input.value;
-
-    expect(ref.current?.value)?.toBe(internalValue);
-  });
-
   it("should clear the value and onClear is triggered", async () => {
     const onClear = jest.fn();
 
@@ -161,7 +157,7 @@ describe("Input", () => {
       />,
     );
 
-    const clearButton = getByRole("button");
+    const clearButton = getByRole("button")!;
 
     expect(clearButton).not.toBeNull();
 
@@ -217,7 +213,7 @@ describe("Input", () => {
       />,
     );
 
-    const clearButton = getByRole("button");
+    const clearButton = getByRole("button")!;
 
     expect(clearButton).not.toBeNull();
 
@@ -276,7 +272,7 @@ describe("Input with React Hook Form", () => {
     input1 = document.querySelector("input[name=withDefaultValue]")!;
     input2 = document.querySelector("input[name=withoutDefaultValue]")!;
     input3 = document.querySelector("input[name=requiredField]")!;
-    submitButton = document.querySelector("button")!;
+    submitButton = document.querySelector('button[type="submit"]')!;
   });
 
   it("should work with defaultValues", () => {