+++ packages/components/popover/src/use-popover.ts
@@ -263,8 +263,7 @@ export function usePopover(originalProps: UsePopoverProps) {
 
       return {
         "data-slot": "trigger",
-        "aria-haspopup": "dialog",
-        ...mergeProps(triggerProps, otherProps),
+        ...mergeProps({"aria-haspopup": "dialog"}, triggerProps, otherProps),
         onPress,
         isDisabled,
         className: slots.trigger({